Refactor initial class scan
This change removes the first pass class reader to allow all class information to built in
a single pass for performance reasons. This change also adds asserts to the context class
which validate that it is not used improperly - eg that the closure has been computed
before the subclass, superclass and interface implementor queries are used.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>