Create Method summaries to know parameters that get instantiated
1. Create a callee-caller relationship by traversing callsite for each method
and adding the callsite to each of the callee's list of callers which is maintained
in the MethodInfo.
The callee-caller relationship is computed after hierarchy context closure
has been computed.

2. Updated CallSite to maintain mapping of argument of the call to the BasicValue.
This information is used by Summarizer to propagate parameter usage information
from callee to caller.

3. Added a list of methods that directly call reflection methods of interest,
eg Class.forName(), Class.newInstance() and ServiceLoader.load()
This list serves as the root set for creating method summary.

4. Created a Summarizer that accepts a set of methods, and traverses the call
graph bottup-up (from callee to caller) propagating parameter usage information at
each call point. It also indetifies if the caller is passing a statically known value
to the callee's parameter and adds that value to the instantiatedTypes of the method.

5. With method summary, the code that handles ClassLoaderService.classForName in
ReferenceAnalyzer is not required.

Signed-off-by: Ashutosh Mehra <mehra.ashutosh@ibm.com>