Fix for some bugs
1. In ReferenceAnalyzer, when handling java/lang/reflect/Method.invoke(),
name of the method being called is not correctly set. The name should be obtained
from the MethodValue.
2. In NonLoadingClassWriter.getCommonSuperClass(), deeperOffset should be
initialized to 0 to avoid negative index when accessing elements in the
'deeper' list.
3. MethodInfo.addCallSite() should replace '.' with '/' in the class name
before creating the CallSite. Without this, the code fails to identify certain
methods as referenced.
Eg in the following code, org.eclipse.Foo.bar() is not marked as referenced.
    Class foo = Class.forName("org.eclipse.Foo");
    Method m = foo.getMethod("bar", ...);
    m.invoke(...);

Signed-off-by: Ashutosh Mehra <mehra.ashutosh@ibm.com>