Handle multiple versions of classes in the application
An application can have multipe versions of the same class. But CallSiteArgumentAnalyzer
didn't have the capability to locate correct MethodNode corresponding to the caller at
a given callsite if the multiple versions of the callers were possible.
To prevent this issue the code was updated to avoid processing multiple versions
of the class. Only the first one was processed, and rest were skipped.
But this is not right approach and can result in failure to identify classfile artifacts
that can be reached by versions of the class that was ignored during processing.

This patch adds the association between callsite and the class source so that
CallSiteArgumentAnalyzer can correctly obtain the MethodNode for the caller at a given
callsite. With this change, multiple versions of the class are correctly handled.

Signed-off-by: Ashutosh Mehra <mehra.ashutosh@ibm.com>